Moved code around
