+++ go-winpmem/api.go
