+++ go-winpmem/assets.go
