+++ go-winpmem/imager.go
