+++ go-winpmem/install.go
