+++ go-winpmem/sparse.go
